requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
executing==2.0.1
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.9
zope.event==5.0
zope.interface==6.1